monitor/requirements.txt
PyQt6==6.7.0
colorlog
opencv-python
psutil
pyzmq
nidaqmx
pyqtgraph